+++ CHANGELOG.md
@@ -6,6 +6,17 @@
 > make sure you follow our [migration guide](https://docs.sentry.io/platforms/react-native/migration/) first.
 <!-- prettier-ignore-end -->
 
+## Unreleased
+
+### Dependencies
+
+- Bump Android SDK from v8.33.0 to v8.34.0 ([#5760](https://github.com/getsentry/sentry-react-native/pull/5760))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#8340)
+  - [diff](https://github.com/getsentry/sentry-java/compare/8.33.0...8.34.0)
+- Bump Android SDK Stubs from v8.33.0 to v8.34.0 ([#5761](https://github.com/getsentry/sentry-react-native/pull/5761))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#8340)
+  - [diff](https://github.com/getsentry/sentry-java/compare/8.33.0...8.34.0)
+
 ## 8.3.0
 
 ### Features