+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Dependencies
 
-- Bump Android SDK Stubs from v8.33.0 to v8.34.0 ([#5765](https://github.com/getsentry/sentry-react-native/pull/5765))
+- Bump Android SDK from v8.33.0 to v8.34.0 ([#5760](https://github.com/getsentry/sentry-react-native/pull/5760))
   - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#8340)
   - [diff](https://github.com/getsentry/sentry-java/compare/8.33.0...8.34.0)